requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
Django==4.2.16
django-cloudinary-storage==0.3.0
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
phonenumbers==8.13.46
pillow==10.4.0
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
